backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
boto3==1.26.137
python-dotenv==0.19.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
openai==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4